First pass at version 4.
