+++ README.md
@@ -5,7 +5,7 @@ Python SEO Analyzer
 
 An SEO tool that analyzes the structure of a site, crawls the site, counts words in the body of the site and warns of any technical SEO issues.
 
-Requires Python 3.4+, BeautifulSoup4 and urllib3.
+Requires Python 3.6+, BeautifulSoup4 and urllib3.
 
 Installation
 ------------
