+++ tests/test_analyzer.py
@@ -1,0 +1,6 @@
+from seoanalyzer import analyze
+
+def test_print_output():
+    output = analyze('https://www.sethserver.com/tests/utf8.html')
+
+    assert False
