+++ tests/test_http.py
@@ -1,0 +1,4 @@
+from seoanalyzer import http
+
+def test_http():
+    assert http.http.get('https://www.sethserver.com/tests/utf8.html')
